+++ cypress/cypress/component/Input/BasicInputField.cy.tsx
@@ -3,6 +3,9 @@ import { Form, Input } from "react-hook-form-components";
 import { faker } from "@faker-js/faker";
 import * as yup from "yup";
 import { yupResolver } from "@hookform/resolvers/yup";
+import { faPencil } from "@fortawesome/free-solid-svg-icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { InputGroupText } from "reactstrap";
 
 describe("Input.cy.tsx", () => {
   it("basic text input works", () => {
@@ -312,5 +315,33 @@ describe("Input.cy.tsx", () => {
 
     cy.contains("label", name).click();
     cy.get(`input[id=${name}]`).getSelectedText().should("eq", value);
+  });
+
+  it("contains pencil icon if provided for Input with no type", () => {
+    const name = faker.random.alpha(10);
+    const schema = yup.object().shape({
+      [name]: yup.date(),
+    });
+
+    cy.mount(
+      <Form
+        onSubmit={() => {
+          // Nothing to do
+        }}
+        resolver={yupResolver(schema)}
+      >
+        <Input
+          name={name}
+          label={name}
+          addonRight={
+            <InputGroupText>
+              <FontAwesomeIcon icon={faPencil} />
+            </InputGroupText>
+          }
+        />
+      </Form>,
+    );
+
+    cy.get(`label[for=${name}]`).parent().find("svg[data-icon=pencil]");
   });
 });