+++ src/lib/FormGroupLayout.tsx
@@ -1,6 +1,6 @@
-import { PropsWithChildren } from "react";
+import { PropsWithChildren, ReactNode } from "react";
 import { FieldError, FieldValues, get, useFormContext } from "react-hook-form";
-import { FormGroup, FormFeedback, FormText } from "reactstrap";
+import { FormGroup, FormFeedback, FormText, InputGroup } from "reactstrap";
 import { useSafeNameId } from "src/lib/hooks/useSafeNameId";
 import { CommonInputProps } from "./types/CommonInputProps";
 import "./styles/FormGroupLayout.css";
@@ -9,10 +9,12 @@ import { FormGroupLayoutLabel } from "./FormGroupLayoutLabel";
 interface FormGroupLayoutProps<T extends FieldValues>
   extends PropsWithChildren<Pick<CommonInputProps<T>, "helpText" | "label" | "name" | "id" | "labelToolTip" | "inputOnly">> {
   layout?: "checkbox" | "switch";
+  addonLeft?: ReactNode;
+  addonRight?: ReactNode;
 }
 
 const FormGroupLayout = <T extends FieldValues>(props: FormGroupLayoutProps<T>) => {
-  const { label, helpText, children, layout, labelToolTip, inputOnly } = props;
+  const { label, helpText, children, layout, labelToolTip, inputOnly, addonLeft, addonRight } = props;
   const { name, id } = useSafeNameId(props.name, props.id);
   const {
     formState: { errors },
@@ -36,7 +38,15 @@ const FormGroupLayout = <T extends FieldValues>(props: FormGroupLayoutProps<T>) 
   ) : (
     <FormGroup switch={switchLayout ? true : undefined} check={checkboxLayout ? true : undefined}>
       <FormGroupLayoutLabel<T> label={label} fieldName={name} fieldId={id} tooltip={labelToolTip} layout={layout} />
-      {children}
+      <InputGroup
+        style={{
+          flexWrap: "nowrap",
+        }}
+      >
+        {addonLeft}
+        {children}
+        {addonRight}
+      </InputGroup>
       <FormFeedback>{errorMessage}</FormFeedback>
       {helpText && <FormText>{helpText}</FormText>}
     </FormGroup>